new eval_gflownet script
